requirements.txt
# Core dependencies
torch>=2.4.0
torchvision
torchaudio
transformers>=4.48.0
accelerate
diffusers>=0.31.0
huggingface_hub
huggingface-cli
numpy
pillow
tqdm
scipy
trimesh
rembg[gpu,onnxruntime]
einops
omegaconf
jaxtyping
typeguard
imageio
imageio-ffmpeg
ninja
pyyaml
rich

# RunPod Serverless
runpod

# 3D and Graphics
pysdf
mcubes
pyrender
opencv-python
